Unified phone mount design
I have adapted my phone mount design to suit the OpenBot. This phone mount design consists of two components that are print-in-place already assembled which serves to simplify the project and provide an increased level of robustness. An elastic band can be used to provide the mechanical grip for greater accessibility and reduced cost. It will be possible to optimise the physical geometry for a smaller foot print and to have a different grip range as required pending feedback. https://www.thingiverse.com/thing:4297510